3.4.0 Refactored AkaniPlayersCommand to AkaniAdminCommand and updated priority values in ScoreboardProvider.

+++ src/main/java/it/einjojo/akani/essentials/scoreboard/defaults/DefaultScoreboardProvider.java
@@ -22,7 +22,7 @@ public class DefaultScoreboardProvider implements ScoreboardProvider {
     }
 
     @Override
-    public int priority() {
+    public short priority() {
         return 0;
     }
